refactor: eliminate ExternalTemplateSearchModal duplication (#397)
- Remove duplicated ExternalTemplateSearchModal implementations from RecipeEditModal and ItemGroupEditModal
- Both components now use the shared ExternalTemplateSearchModal from ~/sections/search/components/
- Move component-specific logic (handleNewItemGroup) to component usage level
- Add missing import for ExternalTemplateSearchModal in ItemGroupEditModal
- Maintain existing functionality while reducing code duplication

Fixes #397
